Use `BigInteger` instead of `decimal` for `Ratio` (#20)
* Use `BigInteger` instead of `decimal` for `Ratio`

* - cleanup

* - use readonly property for one eth

* - make `OneEth` static readonly property

+++ src/TokenSchedule.FluentValidation/Models/IValidatedScheduleItem.cs
@@ -1,10 +1,11 @@
 ﻿using System;
+using System.Numerics;
 
 namespace TokenSchedule.FluentValidation.Models
 {
     public interface IValidatedScheduleItem
     {
-        public decimal Ratio { get; }
+        public BigInteger Ratio { get; }
         public DateTime StartDate { get; }
         public DateTime? FinishDate { get; }
     }
